fix(sublocation): update
- update readme
- refactor updateLocation repository function
- create sublocations for a sublocation
