+++ server/repositories/location/updateLocation.js
@@ -2,6 +2,7 @@ import buildLocationUpdateData from '../../adapters/buildLocationUpdateData'
 import errorHelper from '../../helpers/errorHelper'
 import isValidLocationUpdate from '../validations/isValidLocationUpdate'
 import updateParentLocation from './updateParentLocation'
+import createSubLocations from './subLocations/createSubLocations'
 import getLocationById from './getLocationById'
 
 const updateLocation = async (db, locationId, locationUpdate) => {
@@ -12,6 +13,10 @@ const updateLocation = async (db, locationId, locationUpdate) => {
     if (locationUpdateError.length) {
       const error = errorHelper.createError('400', `${locationToUpdate.name} has sub-locations and cannot be updated directly.`)
       throw (error)
+    }
+
+    if (locationUpdate.subLocations) {
+      await createSubLocations(db, locationId, locationUpdate.subLocations)
     }
 
     const locationUpdateData = buildLocationUpdateData(locationToUpdate, locationUpdate)
